Split huge test file
